+++ packages/ricos-content/web/src/urlValidators/urlValidators.spec.ts
@@ -1,4 +1,4 @@
-import { isValidExactUrl, isValidUrl, isValidTelUrl } from './urlValidators';
+import { isValidExactUrl, isValidUrl, isValidTelUrl, normalizeUrl } from './urlValidators';
 describe('URL Validators', () => {
   it('should verify string with url in it', () => {
     const iframeCode =
@@ -16,5 +16,20 @@ describe('URL Validators', () => {
   it('should validate tel: URL', () => {
     expect(isValidTelUrl('tel:+97254-6449-555')).toBeTruthy();
     expect(isValidTelUrl('https://web.url')).toBeFalsy();
+  });
+
+  it('should validate tel: URL', () => {
+    expect(isValidTelUrl('tel:+97254-6449-555')).toBeTruthy();
+    expect(isValidTelUrl('https://web.url')).toBeFalsy();
+  });
+
+  it('should recgonize emails', () => {
+    const fuzzyTldrMail = 'example@email.gay';
+    expect(isValidUrl(fuzzyTldrMail)).toBe(true);
+  });
+
+  it('should normalize emails', () => {
+    const fuzzyTldrMail = 'example@email.gay';
+    expect(normalizeUrl(fuzzyTldrMail)).toBe('mailto:' + fuzzyTldrMail);
   });
 });